requirements.txt
PyMuPDF==1.20.2
Pillow==9.2.0
fpdf==1.7.2